+++ src/noahtheduke/splint.clj
@@ -38,6 +38,7 @@
     noahtheduke.splint.rules.lint.loop-empty-when
     noahtheduke.splint.rules.lint.missing-body-in-when
     noahtheduke.splint.rules.lint.not-empty
+    noahtheduke.splint.rules.lint.prefer-method-values
     noahtheduke.splint.rules.lint.prefer-require-over-use
     noahtheduke.splint.rules.lint.redundant-call
     noahtheduke.splint.rules.lint.require-explicit-param-tags