+++ src/noahtheduke/splint/rules/performance/dot_equals.clj
@@ -4,6 +4,7 @@
 
 (ns ^:no-doc noahtheduke.splint.rules.performance.dot-equals
   (:require
+    [noahtheduke.splint.diagnostic :refer [->diagnostic]]
     [noahtheduke.splint.rules :refer [defrule]]))
 
 (set! *warn-on-reflection* true)
@@ -13,6 +14,8 @@
 
   Currently only checks string literals.
 
+  If `lint/prefer-method-values` is enabled, then the suggestion will use that syntax.
+
   Examples:
 
   ; bad
@@ -20,9 +23,18 @@
 
   ; good
   (.equals \"foo\" s)
+  (String/equals \"foo\" s)
   "
   {:patterns ['(= (? string string?) ?any)
               '(= ?any (? string string?))]
-   :message "Rely on `.equals` when comparing against string literals."
    :ext :clj
-   :replace '(.equals ?string ?any)})
+   :on-match (fn [ctx rule form {:syms [?string ?any]}]
+               (let [method-values (:enabled ('lint/prefer-method-values (:config ctx)))
+                     replace-form (if method-values
+                                    (list 'String/equals ?string ?any)
+                                    (list '.equals ?string ?any))
+                     msg (if method-values
+                           "Rely on `String/equals` when comparing against string literals."
+                           "Rely on `.equals` when comparing against string literals.")]
+                 (->diagnostic ctx rule form {:replace-form replace-form
+                                              :message msg})))})